# WMS/apps/api-gateway/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IWarehouse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"layout_type": {
						"type": "string"
					},
					"layout_data": {
						"items": {
							"items": {
								"type": "number",
								"format": "double"
							},
							"type": "array"
						},
						"type": "array",
						"nullable": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"code",
					"name",
					"description",
					"width",
					"height",
					"layout_type",
					"layout_data",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateWarehouseDTO": {
				"properties": {
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"layout_type": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"name",
					"width",
					"height",
					"layout_type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateWarehouseDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"layout_type": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SlotType": {
				"enum": [
					"STORAGE",
					"AISLE",
					"BLOCKED",
					"PICKUP",
					"DROPOFF",
					"CHARGING"
				],
				"type": "string"
			},
			"SlotStatus": {
				"enum": [
					"AVAILABLE",
					"OCCUPIED",
					"RESERVED",
					"MAINTENANCE"
				],
				"type": "string"
			},
			"IWarehouseSlot": {
				"properties": {
					"id": {
						"type": "string"
					},
					"warehouse_id": {
						"type": "string"
					},
					"slot_code": {
						"type": "string"
					},
					"x": {
						"type": "number",
						"format": "double"
					},
					"y": {
						"type": "number",
						"format": "double"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"slot_type": {
						"$ref": "#/components/schemas/SlotType"
					},
					"occupied_percent": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/SlotStatus"
					},
					"metadata": {},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"warehouse_id",
					"slot_code",
					"x",
					"y",
					"width",
					"height",
					"slot_type",
					"occupied_percent",
					"status",
					"metadata",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateWarehouseSlotDTO": {
				"properties": {
					"warehouse_id": {
						"type": "string"
					},
					"slot_code": {
						"type": "string"
					},
					"x": {
						"type": "number",
						"format": "double"
					},
					"y": {
						"type": "number",
						"format": "double"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"slot_type": {
						"$ref": "#/components/schemas/SlotType"
					},
					"metadata": {}
				},
				"required": [
					"warehouse_id",
					"slot_code",
					"x",
					"y",
					"width",
					"height",
					"slot_type"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateWarehouseSlotDTO": {
				"properties": {
					"slot_type": {
						"$ref": "#/components/schemas/SlotType"
					},
					"occupied_percent": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/SlotStatus"
					},
					"metadata": {}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IProduct": {
				"properties": {
					"id": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"code",
					"name",
					"description",
					"width",
					"height",
					"weight",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProductDTO": {
				"properties": {
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"code",
					"name",
					"width",
					"height",
					"weight"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProductDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"width": {
						"type": "number",
						"format": "double"
					},
					"height": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"HealthStatus": {
				"description": "Response shape for the health check endpoint",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"healthy",
							"degraded"
						]
					},
					"uptime": {
						"type": "number",
						"format": "double"
					},
					"timestamp": {
						"type": "string"
					},
					"services": {
						"properties": {
							"cloudinary": {
								"type": "string",
								"enum": [
									"configured",
									"not_configured"
								]
							},
							"database": {
								"type": "string",
								"enum": [
									"connected",
									"disconnected"
								]
							},
							"redis": {
								"type": "string",
								"enum": [
									"connected",
									"disconnected"
								]
							}
						},
						"required": [
							"cloudinary",
							"database",
							"redis"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"uptime",
					"timestamp",
					"services"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "wms-backend-root",
		"version": "1.0.0",
		"description": "Root of WMS Backend Monorepo",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/master-data/warehouses": {
			"get": {
				"operationId": "GetAllWarehouses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IWarehouse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateWarehouse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IWarehouse"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWarehouseDTO"
							}
						}
					}
				}
			}
		},
		"/master-data/warehouses/{idOrCode}": {
			"get": {
				"operationId": "GetWarehouseById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IWarehouse"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idOrCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/master-data/warehouses/{id}": {
			"put": {
				"operationId": "UpdateWarehouse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IWarehouse"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWarehouseDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWarehouse",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/master-data/warehouses/{idOrCode}/slots": {
			"get": {
				"operationId": "GetSlotsByWarehouseId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IWarehouseSlot"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idOrCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/master-data/warehouses/{warehouseId}/slots/available": {
			"get": {
				"operationId": "GetAvailableSlots",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IWarehouseSlot"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "warehouseId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/master-data/slots/{id}": {
			"get": {
				"operationId": "GetSlotById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IWarehouseSlot"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSlot",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IWarehouseSlot"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWarehouseSlotDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSlot",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/master-data/slots": {
			"post": {
				"operationId": "CreateSlot",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IWarehouseSlot"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWarehouseSlotDTO"
							}
						}
					}
				}
			}
		},
		"/master-data/products": {
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IProduct"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProduct"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductDTO"
							}
						}
					}
				}
			}
		},
		"/master-data/products/{id}": {
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProduct"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IProduct"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Master Data"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/health": {
			"get": {
				"operationId": "GetHealth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HealthStatus"
								}
							}
						}
					}
				},
				"description": "System health check — verifies connectivity of all core dependencies.\r\nUsed by deployment platforms (Render, Railway, AWS ECS) to determine app status.",
				"tags": [
					"System"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "https://localhost:3000/api"
		}
	]
}